Fixed: - The ability for stopping and starting the schedule. Added: - Added more terminal and window control options And more :)
