+++ src/Core/TerminalScheduleUpdate.java
@@ -1,4 +1,4 @@
-package Terminal;
+package Core;
 
 public interface TerminalScheduleUpdate {
     void Update();
